+++ packages/components/src/tag/docs/tag-group/item/invalid.tsx
@@ -1,0 +1,17 @@
+import { Tag, TagGroup } from "@hopper-ui/components";
+
+export default function Example() {
+    return (
+        <TagGroup aria-label="Jobs">
+            <Tag id="designer" isInvalid>
+                Designer
+            </Tag>
+            <Tag id="developer">
+                Developer
+            </Tag>
+            <Tag id="manager">
+                Manager
+            </Tag>
+        </TagGroup>
+    );
+}